Changed multiread_fixeddss ds_size include key bytes.
Previously multiread_fixeddss was spreading the ds_size bytes over the values
only (actually the calculation for this had a bug, but this was approximately
the effect).

Now ds_size is the budget for keys and values. Keys are still fixed at 30B, and
if ds_size is too small relative to multi_size, then the experiment will report
a warning and skip that particular parameter configuration.